Refactor Rust Implementation
